+++ audiolazy/test/test_poly.py
@@ -31,7 +31,8 @@ from itertools import combinations_with_replacement, combinations
 
 # Audiolazy internal imports
 from ..lazy_poly import Poly, lagrange, resample, x
-from ..lazy_misc import almost_eq, orange, xrange, almost_eq_diff, blocks
+from ..lazy_misc import almost_eq, almost_eq_diff, blocks
+from ..lazy_compat import orange, xrange
 from ..lazy_math import inf
 from ..lazy_filters import z
 from ..lazy_itertools import count